SMACB/Programa.py: movido codigo compartible para hacer otros programas 2
