+++ bin/generaPrograma.py
@@ -1,102 +1,12 @@
 import sys
 
-import numpy as np
 from configargparse import ArgumentParser
-from reportlab.lib import colors
 from reportlab.lib.pagesizes import A4, landscape
-from reportlab.lib.styles import ParagraphStyle, getSampleStyleSheet
 from reportlab.lib.units import mm
-from reportlab.platypus import (Table, SimpleDocTemplate, Paragraph, TableStyle, Spacer, NextPageTemplate, PageTemplate,
-                                Frame, PageBreak)
-from scipy import stats
+from reportlab.platypus import (SimpleDocTemplate, Spacer, NextPageTemplate, PageTemplate, Frame, PageBreak)
 
-import SMACB.Programa
-from SMACB.Programa import auxCalculaBalanceStr, estadsEquipoPortada, listaEquipos, paginasJugadores, \
-    reportTrayectoriaEquipos, tablaLiga, datosRestoJornada
-from SMACB.TemporadaACB import TemporadaACB, precalculaOrdenEstadsLiga, COLSESTADSASCENDING, equipo2clasif
-from Utils.FechaHora import Time2Str
-
-estadGlobales = None
-estadGlobalesOrden = None
-clasifLiga = None
-
-ESTADISTICOEQ = 'mean'
-ESTADISTICOJUG = 'mean'
-ANCHOTIROS = 16
-ANCHOREBOTES = 14
-
-ESTILOS = getSampleStyleSheet()
-
-
-def cabeceraPortada(partido, tempData):
-    datosLocal = partido['equipos']['Local']
-    datosVisit = partido['equipos']['Visitante']
-    compo = partido['cod_competicion']
-    edicion = partido['cod_edicion']
-    j = partido['jornada']
-    fh = Time2Str(partido['fechaPartido'])
-
-    style = ParagraphStyle('cabStyle', align='center', fontName='Helvetica', fontSize=20, leading=22, )
-
-    cadenaCentral = Paragraph(
-        f"<para align='center' fontName='Helvetica' fontSize=20 leading=22><b>{compo}</b> {edicion} - " + f"J: <b>{j}</b><br/>{fh}</para>",
-        style)
-
-    cabLocal = datosCabEquipo(datosLocal, tempData, partido['fechaPartido'])
-    cabVisit = datosCabEquipo(datosVisit, tempData, partido['fechaPartido'])
-
-    tStyle = TableStyle([('BOX', (0, 0), (-1, -1), 2, colors.black), ('VALIGN', (0, 0), (-1, -1), 'MIDDLE'),
-                         ('GRID', (0, 0), (-1, -1), 0.5, colors.black)])
-    t = Table(data=[[cabLocal, cadenaCentral, cabVisit]], colWidths=[60 * mm, 80 * mm, 60 * mm], style=tStyle)  #
-
-    return t
-
-
-def cargaTemporada(fname):
-    result = TemporadaACB()
-    result.cargaTemporada(fname)
-
-    return result
-
-
-def datosCabEquipo(datosEq, tempData, fecha):
-    recuperaClasifLiga(tempData, fecha)
-
-    # TODO: Imagen (descargar imagen de escudo y plantarla)
-    nombre = datosEq['nombcorto']
-
-    clasifAux = equipo2clasif(clasifLiga, datosEq['abrev'])
-    clasifStr = auxCalculaBalanceStr(clasifAux)
-
-    result = [Paragraph(f"<para align='center' fontSize='16' leading='17'><b>{nombre}</b></para>"),
-              Paragraph(f"<para align='center' fontSize='14'>{clasifStr}</para>")]
-
-    return result
-
-
-def recuperaEstadsGlobales(tempData):
-    global estadGlobales, estadGlobalesOrden
-
-    if estadGlobales is None:
-        estadGlobales = tempData.dfEstadsLiga()
-        estadGlobalesOrden = precalculaOrdenEstadsLiga(estadGlobales, COLSESTADSASCENDING)
-
-
-def recuperaClasifLiga(tempData, fecha=None):
-    global clasifLiga
-
-    print(tempData)
-    if clasifLiga is None:
-        clasifLiga = tempData.clasifLiga(fecha)
-        jugados = np.array([eq['Jug'] for eq in clasifLiga])
-        modaJug = stats.mode(jugados, keepdims=False).mode
-
-        for eq in clasifLiga:
-            if eq['Jug'] != modaJug:
-                pendientes = modaJug - eq['Jug']
-                aux = "*" if (abs(pendientes) == 1) else pendientes
-
-                eq.update({'pendientes': aux})
+from SMACB.Programa import estadsEquipoPortada, listaEquipos, paginasJugadores, reportTrayectoriaEquipos, tablaLiga, \
+    cabeceraPortada, cargaTemporada
 
 
 def preparaLibro(outfile, tempData, datosSig):
@@ -119,8 +29,6 @@ def preparaLibro(outfile, tempData, datosSig):
     story = []
 
     (sigPartido, abrEqs, juIzda, peIzda, juDcha, peDcha, targLocal) = datosSig
-    datosJor = datosRestoJornada(tempData, datosSig)
-    exit(1)
 
     antecedentes = {p.url for p in juIzda}.intersection({p.url for p in juDcha})
 
@@ -175,7 +83,7 @@ def parse_arguments():
 def main(args):
     tempData = cargaTemporada(args.acbfile)
 
-    if SMACB.Programa.listaEquipos:
+    if args.listaEquipos:
         listaEquipos(tempData, args.quiet)
 
     REQARGS = ['equipo', 'outfile']
